requirements.txt
svglib
Pillow
urllib
requests
reportlab